+++ sbin/fsck_ffs/pass2.c
@@ -137,8 +137,8 @@ pass2()
 	for (inpp = inpsort; inpp < inpend; inpp++) {
 		if (got_siginfo) {
 			printf("%s: phase 2: dir %d of %d (%d%%)\n", cdevname,
-			    inpp - inpsort, inplast, (inpp - inpsort) * 100 /
-			    inplast);
+			    inpp - inpsort, (int)inplast,
+			    (int)((inpp - inpsort) * 100 / inplast));
 			got_siginfo = 0;
 		}
 		inp = *inpp;
